final commit

+++ src/app/header/header.component.ts
@@ -3,6 +3,7 @@ import {DataProductService} from "../services/data.product.service";
 import {FamilyProduct} from "../classes/FamilyProduct";
 import {AppComponent} from "../app.component";
 import {CartLine} from "../classes/CartLine";
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-header',
@@ -27,7 +28,25 @@ export class HeaderComponent implements OnInit {
   }
 
   deleteLine(line: CartLine) {
-    this.appComponent.cart.deleteLine(line);
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "estas seguro que deseas borrar el producto " + line.product.name + "?",
+      icon: 'error',
+      showCancelButton: true,
+      confirmButtonColor: '#00B0F0',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Borrarlo!',
+      cancelButtonText: 'No'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.appComponent.cart.deleteLine(line);
+        Swal.fire(
+          'Borrado!',
+          'El producto se borro de tu pedido',
+          'success'
+        )
+      }
+    })
   }
 
 }
